// sigpcb-backoffice-frontend/src/app/core/interfaces/date.ts
export type FullMonth =
  | 'Janvier'
  | 'Février'
  | 'Mars'
  | 'Avril'
  | 'Mai'
  | 'Juin'
  | 'Juillet'
  | 'Août'
  | 'Septembre'
  | 'Octobre'
  | 'Novembre'
  | 'Décembre';

export type FullDay =
  | 'Lundi'
  | 'Mardi'
  | 'Mercredi'
  | 'Jeudi'
  | 'Vendredi'
  | 'Samedi'
  | 'Dimanche';